Updates Installer, adds TaskCreateUserOne, and makes work with tests
@jimsafley I think this gets close to what you need. It's possible that the install schema hasn't kept up with latest work. I merge develop a day or so ago, but might have missed newest, or changes to branch you are using.

/install_data/write_schema_sql.php is there to update the install sql based on what the entity manager says the current schema should be, so until we have migrations going that might be something to run to keep the schema up to date with latest code.

I've broken separate tasks for installation into the stuff in the Install directory. Probably needs to be broken into more subdirectories before merging into develop. It might also be worthwhile for me to make the errors and messages follow the pattern you've developed.

The Bootstrap for testing has static methods for installing and dropping the tables. The UserOneTest drops everything and installs the tables anew.

application.config.php.dist has an array of install_tasks that lists out the names of classes for installation tasks to do. The idea is that we can build this up as we need new things to install.

It sounds like the next one to build is \Omeka\Install\DublinCore. If we can find some pair-programming time on that among the entire dev team, I think that'd be helpful learning all the way around.