A stab at abstracting some of the JS for the sidebar.
Creates a closure returning a function to handle sidebar actions generically. Data from the clicked HTML element is stored in the closure's scope, so the returned function just uses it.

Possible improvement in the HTML based on this would be to use a class 'sidebar-action' to attach the same handler more generally. (i.e., 'sidebar-details', 'sidebar-delete', 'sidebar-close' are redundant to the JS in this system.

Doesn't yet address how things work in the item-add branch.