Separate page content from page "shell"
This way, the content renders first and can set variables in the "shell"
to alter its display. As an example, this change also adds a
"displayNavigation" variable which the blocks can alter to make the page
not show navigation if a certain block is present or other condition is
met.