Pass page to block layout form method
Before, BlockLayoutInterface::form() only took the site object and block object (if available). When the block argument was null there was no way to derive the page object for layouts that need it. In this case, when adding a PageTitle block to a page, it had no way to know the page title.