Consolidate standard "nice link to resource" code
Browses, sidebar-selects, item sidebars and more all basically have (or
should have) the same pattern for linking to resources by showing the
thumbnail and a link. Due to clumsiness of the necessary code, each does
this separately, and several areas (the browses) use non-optimal markup
separating the images from the links because it's easier to do.

This change adds a new linkPretty() helper for Resource enttiy
representations that codifies the "standard" link for a resource. As
part of this, it adds a standard class for the link itself so it's
easier to target, allowing the removal of many separate areas applying
essentially the same styling.