Add "forked" doctrine AbstractProxyFactory
S's proxy loading for modules has always relied on autoloading: this
has reliably worked with Doctrine but never been officially allowed.
S has used a hack of setting the autogenerate strategy mode to one
that doesn't exist. Recent versions of doctrine/common now check that
the mode is a real one so this no longer works, and the older usable
version of common instead contains a bug that accidentally forces the
mode to one of only 2 options...

This works around the problem by loading our own copy of the
AbstractProxyFactory which includes an autogenerate "mode" that
corresponds to the value S's hack has used.