Be much more aggressive in EM reset during batches
Instead of detaching the touched entities, keep track of what the
EM was managing before the batch, and detach everything not in that set.

Intended to resolve an issue where Criteria usage causes entities to be
managed by the EM but not connected to the related entity the Criteria
was run against, meaning our usual strategy of relying on "cascade
detach" did not work.

Fix #1415