Escaped reserved word "value" (#1437)
"Value" is a reserved word in some versions of MySQL/MariaDB. If you're using one of those versions, queries that access or modify the "value" table currently fail. Escaping "value" as a table and column name fixes the error for people running one of those versions, and should not affect usage by people not running one of those versions.