Use JSON to submit values for resources
This is an in-between effort to fix issues users are reporting with very
large items with lots of values: as each value is 5 or so individual PHP
array entries, it's not _hugely_ difficult to get to the PHP "input var"
limit, defaulting to 1024. Using JSON gets around this issue as we can
fit many values into what counts as a single "var" for that limit.

Rather than completely replacing the submission process with something
JSON-based, this commit replaces only the handling of the values
themselves. Because of their highly-dynamic nature, the values are
pretty much already handled in a way that maps nicely to serializing to
JSON, and lots of values seems to be the main way users are managing to
hit the limit.

Leaving the rest of the form in place as-is prevents issues with
backwards compatibility of modules that add to the resource add/edit
forms.

The controller side of this implementation is quite conservative: if it
sees no data or an issue parsing the JSON, it throws immediately. The
idea is to make sure that any Javascript issues preventing the values
from being serialized and sent along merely result in errors, and not in
the server-side simply not seeing any values and carrying on
obliviously.

(fix #1330)