Fix error for value pointing to private items
The code preventing this kind of error was keyed specifically to the
name "resource" for the data type of the value, but we've since
introduced other more specific resource types for values that can point
only to items, sets, etc.

Rather than base this check on the name of the type, this instead
checks if the type class extends from the abstract class for resource
data types. This introduces a new method isHidden to the value
representation to simplify this check.

(fix #1507)

(cherry picked from commit 61abd653da049fd700a0ebf1ed7890dfb4601b17)