Avoid JS errors with interesting fragments
- The main change: switch from using jQuery's lookup to just
getElementById, so there's no question of trying to interpret the
fragment as some odd selector. Should also be incrementally more
efficient for all that matters.
- The second, we already have a body class for not *setting* the
section names as fragments, so let's just look at that for inspecting
the fragment as well: if the body has the class "no-section-hashes" then
we just won't look at the fragment at all.

(fix #1582)