Don't detach entities in in-process batch edit
The detach process used in batch edit/create is pretty brittle,
especially in "synchronous" or in-process contexts where many more
entities are loaded before the batch edit begins, so the results are
less predictable. Typical "bad" results are Doctrine errors about
duplicates or encountering "new" un-persisted entities through a
relationship.

The detach also serves very little purpose in this
context, as it's only there to improve performance _between_ multiple
batches, and the in-process edits only do a single batch.

This commit adds a new request option for batch edits and creates,
"detachEntity", defaulted to true, and has the controllers for the
in-process batch updates set it to false.

(#1690)