Pass used and/or site id parameters to the property search query
if the options to limit the scope of the properties offered in
the property select element is set.

This will also filter the class, so class can be filtered even
if the user is using the more strict 'restrict to template' filter
to limit properties. Properties, however, get their value options
from PropertySelect if the user has selected 'restrict to template'