Initial PHP 8.2 readiness
Mostly this is work eliminating dynamically-created properties, which
are deprecated in 8.2. It also involves some Laminas updates to the same
purpose.

Much of this happens in tests, though there are several instances in the
application code.

On tests, also moving Model/Entity to just Entity to match the application
structure.