After login, redirect to admin only if user is allowed
By default all roles have access to administration dashboard but modules
can change permissions or create new roles that forbid access to it.
Users with that kind of role should not be redirected to the admin
dashboard, as they will land on an error page (with a confusing
"Successfully logged in" message)

This patch verify that the user is allowed to see the admin dashboard
before redirecting to it. If the user is not allowed, it will be
redirected to the front page