Explicitly cache identity lookups from Doctrine
This fixes a performance regression from
613c45c7af285b9194f5557f828fba9d5dbbdc57 (also backported to 3.2.3)

Previously we relied on Doctrine's inherent ID lookup caching to avoid
redundant queries for the current user, but adding the query requirement
to also check if the user was active resulted in Doctrine always
performing the query, leading to large numbers of identical queries for
the current user.

To avoid this, we store the result of the first lookup against Doctrine
and return its result for all later requests for the user identity
within the same request.