Explicitly require doctrine/annotations
Doctrine's now looser about requiring this library. We didn't notice,
because one of our dev dependencies does still require it, with the
result that installs from git checkouts were fine but "production"
zip builds were not. The fix is just to explicitly require the library
as a regular dependency.

As a side effect, we're also updating it to 1.14.1.