Allow fulltext search sorting (#1973)
This is done by passing a sort_by_default flag on queries that use
the default sort. The fulltext search only sorts on relevance when
this flag is present. A consequence is that the user must now
explicitly pass sort_by_default on an API query to sort by relevance.