Fix browse views and usage of lang filtering
This fixes two related issues: first, a regression from 4.0 for the
default browse views: when moving them to use linkPretty to avoid
redundant links, the logic for choosing different title properties and
langauges was broken. This keeps the linkPretty markup structure but
restores the browse heading property and lang filtering features.

Second, this harmonizes treatment of filter_locale_values across the
site views. Throughout, all usages pass the configured lang plus the
empty string which allows values with unset langs to pass the filter;
this matches with the pre-4.0 browse page behavior and the previous and
continuing behavior of the resource-values view.

Together, these also result in fixing media show to respect
filter_locale_values, and fixing the browse-preview block markup to
properly match the normal browse markup.

(fix #1996)