Remove check for version for enabling DB sessions
The version being checked against is an ancient pre-1.0 release that you
can't upgrade straight to current versions anyway. The check here
therefore was acting in practice as a slightly circuitous check that the
database was installed and accessible (and relied on a deprecated check
against null to accomplish it).

Instead of that, just skip session setup when we detect that we aren't
installed, and use the DB-backed sessions otherwise (or the configured
backend).