Deny access to user edit action for low-priv users
Previously you could open the user edit page but not successfully make
changes if you didn't have permission to edit that user. Now you won't
be able to view that page at all.

The edit page shows that user's settings and the "identity" half of
their API keys, which shouldn't be accessible.