Render non-browser-native images in gallery
We'll include these in the gallery when we have thumbnails for them and
they're of an image/* mimetype, but they're not in the whitelist as
being browser-supported; these will get rendered by showing the "large"
thumbnail image.

Alongside this there's some cleanup/refactoring to the lightgallery
helper, and a change to the gallery block for media to make it only use
the gallery when the gallery supports the media being viewed, otherwise
falling back to the standard embed.

(fix #2068)