Allow more log paths; catch instead of test
Removing checks on is_writable and is_file allows setting the log path
to a stream URL (for example, php://stdout). To maintain the same
behavior as now, we catch the exception raised by the log stream writer
constructor when the stream isn't writable (and as a bonus log the issue
to the PHP error log).