Reduce redundancy in screen reader labels.
This looks to reduce repetitive screen reader "search" announcements on
the search forms. By providing the form with the "search" role, we can
trust the user to know when they are in a search context. The previous
approach using "aria-labelledby" was creating a lot of extra readings of
element labels in NVDA, JAWS, and Voiceover.

Other adjustments:
* In the admin, changing resource type no longer changes the aria label.
* No more changing aria-labels with aria-labelledby as it causes issues
  in localization.

(#2280)